backend/requirements.txt
# =============================================
# CORE
# =============================================
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
django-redis==5.4.0

# =============================================
# DATABASE
# =============================================
psycopg2-binary==2.9.9

# =============================================
# CELERY
# =============================================
celery==5.3.6
redis==5.0.1

# =============================================
# PARSING
# =============================================
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
feedparser==6.0.11

# =============================================
# ML / NLP
# =============================================
tensorflow==2.15.0
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4

# =============================================
# SERVER
# =============================================
gunicorn==21.2.0
whitenoise==6.6.0

# =============================================
# UTILS
# =============================================
python-decouple==3.8
python-dateutil==2.9.0

# =============================================
# DEV / TEST
# =============================================
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-env==1.1.3
factory-boy==3.3.0
Faker==24.3.0

# =============================================
# LINTING
# =============================================
black==24.2.0
flake8==7.0.0
isort==5.13.2
